+++ Assets/Core/Runtime/Scripts/DataBind/Bind/PropertiesBind/FloatPropertiesBind.cs
@@ -2,7 +2,15 @@ using UnityEngine;
 
 namespace KKSFramework.DataBind
 {
-    public class FloatPropertiesBind : BaseValueBindableProperties<Component, float>
+    public class FloatPropertiesBind : BaseValueBindableProperties<Component, float[]>
     {
+        protected override void SetDelegate (float[] values)
+        {
+            targetComponents.ZipForEach (values, (comp, value) =>
+            {
+                var target = comp.GetComponent (targetComponent.GetType ());
+                targetComponent.GetType ().GetProperty (propertyName)?.SetValue (target, value);
+            });
+        }
     }
 }