+++ Assets/Core/Runtime/Scripts/DataBind/Bind/PropertiesBind/ImageSpritePropertiesBind.cs
@@ -4,10 +4,12 @@ using UnityEngine.UI;
 
 namespace KKSFramework.DataBind
 {
-    public class ImageSpritePropertiesBind : BindableProperties<Image, Sprite>
+    public class ImageSpritePropertiesBind : BindableProperties<Image, Sprite[]>
     {
-        protected override Sprite GetDelegate () => targetComponents.First ().GetComponent<Image> ().sprite;
+        protected override Sprite[] GetDelegate () =>
+            targetComponents.Select (x => x.GetComponent<Image> ().sprite).ToArray ();
 
-        protected override void SetDelegate (Sprite value) => targetComponents.Foreach (x => x.sprite = value);
+        protected override void SetDelegate (Sprite[] values) =>
+            targetComponents.ZipForEach (values, (comp, value) => comp.GetComponent<Image> ().sprite = value);
     }
 }