+++ Assets/Core/Runtime/Scripts/DataBind/Bind/PropertiesBind/TextLabelPropertiesBind.cs
@@ -3,10 +3,12 @@ using UnityEngine.UI;
 
 namespace KKSFramework.DataBind
 {
-    public class TextLabelPropertiesBind : BindableProperties<Text, string>
+    public class TextLabelPropertiesBind : BindableProperties<Text, string[]>
     {
-        protected override string GetDelegate () => targetComponents.First ().GetComponent<Text> ().text;
+        protected override string[] GetDelegate () =>
+            targetComponents.Select (x => x.GetComponent<Text> ().text).ToArray ();
 
-        protected override void SetDelegate (string value) => targetComponents.Foreach (x => x.text = value);
+        protected override void SetDelegate (string[] values) =>
+            targetComponents.ZipForEach (values, (comp, value) => { comp.text = value; });
     }
 }